Fix web clients not getting updated about the state of XP evokers
Whenever an inventory item's item_def changes, the game sends an update
to the web client(s). Because information about XP evokers is stored
outside of item_defs, there are no updates when XP evokers recharge
or lose charges.

Fix this by caching the number of charges in the item.plus field
(similar to how wand charges are stored) and sending updates when
necessary.