Add an in-game configuration menu for the consumables panel
This new menu allows changing the panel's orientation, scale, and font
size in-game. Also, it allows minimizing the panel into a "+",
similar to how the chat window can be minimized.

All the changes are done on the client side and not saved between
sessions.

Also, this commit splits the panel's font option into font_size and
font_family, so it is similar to other existing font options.